Make email template translatable
Also make the template more smilar to the other default email templates.

This commit also updates the (non yet translated) german translation file.
That files has a lot of changes, but it's the output of `make localegen`
and I think it makes sense to use that output as a basis for the actual
translation.